level browser layer player counts

+++ server/game/src/data/types/mod.rs
@@ -5,7 +5,7 @@ pub mod game;
 pub mod gd;
 pub mod misc;
 
-use std::sync::atomic::AtomicI64;
+use std::sync::atomic::{AtomicI32, AtomicI64};
 
 pub use audio_frame::*;
 pub use cocos::*;
@@ -14,8 +14,8 @@ pub use esp::types::*;
 pub use game::*;
 pub use gd::*;
 pub use misc::*;
-pub type LevelId = i64;
-pub type AtomicLevelId = AtomicI64;
+pub type LevelId = i32;
+pub type AtomicLevelId = AtomicI32;
 
 pub const fn is_editorcollab_level(id: LevelId) -> bool {
     id > (2 as LevelId).pow(32)
